Fix route imports for modular server package
