+++ client/src/pages/CardsPage.tsx
@@ -17,8 +17,10 @@ import { ImportCardDialog } from "@/components/cards/ImportCardDialog"
 import { useToast } from "@/components/ui/use-toast"
 import { apiFetch } from "@/lib/api-client"
 
-import { useCards, useCard, useDeleteCard, useCardCatalog } from "@/hooks/useCards"
+import { useApplyForCard, useApproveApplication, useCards, useCard, useDeleteCard, useCardCatalog } from "@/hooks/useCards"
+import { useRewardsEstimate } from "@/hooks/useRewards"
 import type { CardRow as CardRowType, CreditCardProduct } from "@/types/api"
+import { AlertTriangle, Loader2 } from "lucide-react"
 
 const currency0 = new Intl.NumberFormat(undefined, { style: "currency", currency: "USD", maximumFractionDigits: 0 })
 const currency2 = new Intl.NumberFormat(undefined, { style: "currency", currency: "USD", maximumFractionDigits: 2 })
@@ -41,27 +43,50 @@ type AnnualFeeFilter = typeof ANNUAL_FEE_FILTERS[number]["value"]
 
 const PAGE_SIZE = 4
 
+const STATUS_FILTERS = [
+    { value: "all", label: "All" },
+    { value: "applied", label: "Applied only" },
+    { value: "not_applied", label: "Not applied" },
+] as const
+type StatusFilter = typeof STATUS_FILTERS[number]["value"]
+
 export default function CardsPage() {
     const { toast } = useToast()
 
     /* =============== LINKED CARDS =============== */
     const cardsQuery = useCards()
     const cards = cardsQuery.data ?? []
+    const [linkedFilter, setLinkedFilter] = useState<StatusFilter>("all")
     const [selectedId, setSelectedId] = useState<string | undefined>()
     const cardDetails = useCard(selectedId)
+    const selectedCardSlug = cardDetails.data?.cardProductSlug ?? null
+    const rewardsEstimate = useRewardsEstimate(selectedCardSlug, { enabled: Boolean(selectedCardSlug) })
 
     const deleteCard = useDeleteCard({
         onSuccess: () => toast({ title: "Card removed", description: "We’ll tidy up your stats." }),
         onError: (error) => toast({ title: "Unable to remove card", description: error.message }),
     })
+    const applyMutation = useApplyForCard()
+    const approveMutation = useApproveApplication()
+    const [pendingSlug, setPendingSlug] = useState<string | null>(null)
+
+    const filteredLinkedCards = useMemo(() => {
+        if (linkedFilter === "all") return cards
+        return cards.filter((card) => {
+            const applied = isCardApplied(card)
+            return linkedFilter === "applied" ? applied : !applied
+        })
+    }, [cards, linkedFilter])
 
     useEffect(() => {
-        if (!cards.length) {
+        if (!filteredLinkedCards.length) {
             setSelectedId(undefined)
             return
         }
-        if (!selectedId || !cards.some((c) => c.id === selectedId)) setSelectedId(cards[0].id)
-    }, [cards, selectedId])
+        if (!selectedId || !filteredLinkedCards.some((c) => c.id === selectedId)) {
+            setSelectedId(filteredLinkedCards[0].id)
+        }
+    }, [filteredLinkedCards, selectedId])
 
     const [dialogOpen, setDialogOpen] = useState(false)
     const [editDialogOpen, setEditDialogOpen] = useState(false)
@@ -72,6 +97,20 @@ export default function CardsPage() {
 
     const summary = cardDetails.data?.summary
     const donutData = useMemo(() => summary?.byCategory ?? [], [summary])
+    const showLinkingBanner = cardDetails.data ? isCardApplied(cardDetails.data) && !cardDetails.data.mask : false
+    const rewardsData = rewardsEstimate.data
+    const rewardsByCategory = useMemo(
+        () => Object.entries(rewardsData?.earnings.byCategory ?? {}).sort((a, b) => Number(b[1]) - Number(a[1])),
+        [rewardsData]
+    )
+    const rewardsProjection = rewardsData ? currency2.format(rewardsData.projectedMonthly) : currency2.format(0)
+    const statusCaption = cardDetails.data
+        ? isCardApplied(cardDetails.data) && cardDetails.data.appliedAt
+            ? `Applied ${new Date(cardDetails.data.appliedAt).toLocaleDateString()}`
+            : cardDetails.data.lastSynced
+                ? `Synced ${new Date(cardDetails.data.lastSynced).toLocaleDateString()}`
+                : undefined
+        : undefined
 
     const handleDelete = (id: string) => deleteCard.mutate(id)
     const handleEdit = (id: string) => {
@@ -117,44 +156,74 @@ export default function CardsPage() {
     const [issuerFilter, setIssuerFilter] = useState<string>("all")
     const [categoryFilter, setCategoryFilter] = useState<string>("all")
     const [annualFeeFilter, setAnnualFeeFilter] = useState<AnnualFeeFilter>("all")
+    const [catalogStatusFilter, setCatalogStatusFilter] = useState<StatusFilter>("all")
 
     const [appliedSlugs, setAppliedSlugs] = useState<Set<string>>(new Set())
 
-    const filteredCatalog = useMemo(() => {
-        return catalogCards.filter((card) => {
-            const matchesIssuer = issuerFilter === "all" || card.issuer === issuerFilter
-            const matchesCategory =
-                categoryFilter === "all" || (card.rewards ?? []).some((r) => r.category === categoryFilter)
-            const matchesFee = matchesAnnualFee(card.annual_fee, annualFeeFilter)
-            return matchesIssuer && matchesCategory && matchesFee
+    const catalogWithApplied = useMemo<Array<{ product: CreditCardProduct; applied: boolean }>>(() => {
+        return catalogCards.map((product) => {
+            const hasLinkedSlug = product.slug ? appliedSlugs.has(product.slug) : false
+            const applied = hasLinkedSlug || appliedMatcher(product)
+            return { product, applied }
         })
-    }, [catalogCards, issuerFilter, categoryFilter, annualFeeFilter])
+    }, [catalogCards, appliedMatcher, appliedSlugs])
 
-    // pagination
+    const filteredCatalogEntries = useMemo(() => {
+        return catalogWithApplied.filter(({ product, applied }) => {
+            const matchesIssuer = issuerFilter === "all" || product.issuer === issuerFilter
+            const matchesCategory =
+                categoryFilter === "all" || (product.rewards ?? []).some((r) => r.category === categoryFilter)
+            const matchesFee = matchesAnnualFee(product.annual_fee, annualFeeFilter)
+            const matchesStatus =
+                catalogStatusFilter === "all"
+                    ? true
+                    : catalogStatusFilter === "applied"
+                        ? applied
+                        : !applied
+            return matchesIssuer && matchesCategory && matchesFee && matchesStatus
+        })
+    }, [catalogWithApplied, issuerFilter, categoryFilter, annualFeeFilter, catalogStatusFilter])
+
     const [page, setPage] = useState(1)
     useEffect(() => {
-        setPage(1) // reset to first page on filter changes
-    }, [issuerFilter, categoryFilter, annualFeeFilter, catalogCards.length])
+        setPage(1)
+    }, [issuerFilter, categoryFilter, annualFeeFilter, catalogStatusFilter, filteredCatalogEntries.length])
 
-    const totalPages = Math.max(1, Math.ceil(filteredCatalog.length / PAGE_SIZE))
+    const totalPages = Math.max(1, Math.ceil(filteredCatalogEntries.length / PAGE_SIZE))
     const start = (page - 1) * PAGE_SIZE
     const end = start + PAGE_SIZE
-    const pageItems = filteredCatalog.slice(start, end)
+    const pageItems = filteredCatalogEntries.slice(start, end)
 
     const [activeTab, setActiveTab] = useState<CardsTab>("linked")
 
     const onApply = async (product: CreditCardProduct) => {
+        if (!product.slug) {
+            toast({ title: "Couldn’t start application", description: "Product slug missing." })
+            return
+        }
         try {
-            await apiFetch("/applications", {
-                method: "POST",
-                body: { slug: product.slug, product_name: product.product_name, issuer: product.issuer },
+            setPendingSlug(product.slug)
+            const response = await applyMutation.mutateAsync({ slug: product.slug })
+            setAppliedSlugs((prev) => {
+                const next = new Set(prev)
+                next.add(product.slug!)
+                return next
             })
-            const next = new Set(appliedSlugs)
-            if (product.slug) next.add(product.slug)
-            setAppliedSlugs(next)
             toast({ title: "Application started", description: `${product.product_name} by ${product.issuer}` })
-        } catch (e: any) {
-            toast({ title: "Couldn’t start application", description: e?.message ?? "Unknown error" })
+            try {
+                await approveMutation.mutateAsync({ application_id: response.id })
+                setActiveTab("linked")
+                toast({ title: "Application approved", description: "We added the card to your linked list." })
+            } catch (error: any) {
+                toast({
+                    title: "Approval pending",
+                    description: error?.message ?? "We started your application but couldn't auto-link it.",
+                })
+            }
+        } catch (error: any) {
+            toast({ title: "Couldn’t start application", description: error?.message ?? "Unknown error" })
+        } finally {
+            setPendingSlug(null)
         }
     }
 
@@ -188,7 +257,7 @@ export default function CardsPage() {
                     <div className="flex flex-col gap-6 md:flex-row">
                         <div className="md:w-5/12 space-y-4">
                             <CardSelector
-                                cards={cards}
+                                cards={filteredLinkedCards}
                                 selectedId={selectedId}
                                 onSelect={setSelectedId}
                                 onDelete={handleDelete}
@@ -196,6 +265,8 @@ export default function CardsPage() {
                                 onAdd={() => setDialogOpen(true)}
                                 isLoading={cardsQuery.isLoading}
                                 heightClass="max-h-[780px]"
+                                headerExtras={<StatusFilterPills value={linkedFilter} onChange={setLinkedFilter} />}
+                                emptyMessage={cards.length ? "No cards match this filter." : "No cards linked yet."}
                             />
                         </div>
 
@@ -210,19 +281,90 @@ export default function CardsPage() {
                                 <>
                                     <CreditCardDisplay card={cardDetails.data} />
 
+                                    {showLinkingBanner ? (
+                                        <div className="flex flex-col gap-3 rounded-2xl border border-amber-200 bg-amber-50/80 p-4 text-amber-900 sm:flex-row sm:items-center sm:justify-between">
+                                            <div className="flex items-start gap-3">
+                                                <AlertTriangle className="mt-0.5 h-5 w-5 text-amber-500" />
+                                                <div>
+                                                    <p className="text-sm font-semibold">Finish linking to sync spend</p>
+                                                    <p className="text-xs text-amber-800/80">
+                                                        Add your card details so we can pull transactions automatically.
+                                                    </p>
+                                                </div>
+                                            </div>
+                                            <Button
+                                                variant="outline"
+                                                size="sm"
+                                                className="border-amber-300 text-amber-900 hover:bg-amber-100"
+                                                onClick={() => {
+                                                    setEditingCard(cardDetails.data as CardRowType)
+                                                    setEditDialogOpen(true)
+                                                }}
+                                            >
+                                                Update card
+                                            </Button>
+                                        </div>
+                                    ) : null}
+
                                     <div className="grid grid-cols-1 gap-4 sm:grid-cols-3">
                                         <StatTile label="30-day spend" value={currency2.format(summary?.spend ?? 0)} />
                                         <StatTile label="Transactions" value={(summary?.txns ?? 0).toLocaleString()} />
                                         <StatTile
                                             label="Status"
                                             value={cardDetails.data.status}
-                                            caption={
-                                                cardDetails.data.lastSynced
-                                                    ? `Synced ${new Date(cardDetails.data.lastSynced).toLocaleDateString()}`
-                                                    : undefined
-                                            }
+                                            caption={statusCaption}
                                         />
                                     </div>
+
+                                    {selectedCardSlug ? (
+                                        <Card className="rounded-3xl">
+                                            <CardHeader>
+                                                <CardTitle className="text-lg font-semibold">Cash-back (last 30 days)</CardTitle>
+                                                <CardDescription>
+                                                    {rewardsEstimate.isLoading
+                                                        ? "Crunching the numbers…"
+                                                        : `Projected monthly ${rewardsProjection}`}
+                                                </CardDescription>
+                                            </CardHeader>
+                                            <CardContent className="space-y-3">
+                                                {rewardsEstimate.isLoading ? (
+                                                    <div className="flex items-center gap-2 text-sm text-muted-foreground">
+                                                        <Loader2 className="h-4 w-4 animate-spin" />
+                                                        Calculating rewards…
+                                                    </div>
+                                                ) : rewardsData ? (
+                                                    <>
+                                                        <div className="text-2xl font-semibold">
+                                                            {currency2.format(rewardsData.earnings.total)}
+                                                        </div>
+                                                        {rewardsByCategory.length ? (
+                                                            <div className="grid gap-2 text-sm text-muted-foreground sm:grid-cols-2">
+                                                                {rewardsByCategory.slice(0, 4).map(([category, total]) => (
+                                                                    <div
+                                                                        key={category}
+                                                                        className="flex items-center justify-between rounded-xl bg-muted/60 px-3 py-2"
+                                                                    >
+                                                                        <span className="font-medium text-foreground/80">{category}</span>
+                                                                        <span className="font-semibold text-foreground">
+                                                                            {currency2.format(Number(total))}
+                                                                        </span>
+                                                                    </div>
+                                                                ))}
+                                                            </div>
+                                                        ) : (
+                                                            <p className="text-sm text-muted-foreground">
+                                                                We haven’t detected any category bonuses yet.
+                                                            </p>
+                                                        )}
+                                                    </>
+                                                ) : (
+                                                    <p className="text-sm text-muted-foreground">
+                                                        We don’t have enough spend yet to estimate cash-back for this card.
+                                                    </p>
+                                                )}
+                                            </CardContent>
+                                        </Card>
+                                    ) : null}
 
                                     <Card className="rounded-3xl">
                                         <CardHeader>
@@ -303,8 +445,8 @@ export default function CardsPage() {
                                     <CardDescription>Deterministic catalog data — no PII, no surprises.</CardDescription>
                                 </div>
                                 <span className="text-xs text-muted-foreground">
-                  {catalogQuery.isLoading ? "…" : `${filteredCatalog.length} items`}
-                </span>
+                                    {catalogQuery.isLoading ? "…" : `${filteredCatalogEntries.length} items`}
+                                </span>
                             </div>
                         </CardHeader>
                         <CardContent className="space-y-4 p-6 md:p-8 pt-4">
@@ -350,6 +492,9 @@ export default function CardsPage() {
                                     </SelectContent>
                                 </Select>
                             </div>
+                            <div className="flex flex-wrap items-center justify-end gap-2 pt-1">
+                                <StatusFilterPills value={catalogStatusFilter} onChange={setCatalogStatusFilter} />
+                            </div>
                         </CardContent>
                     </Card>
 
@@ -359,7 +504,7 @@ export default function CardsPage() {
                                 Loading catalog…
                             </CardContent>
                         </Card>
-                    ) : filteredCatalog.length === 0 ? (
+                    ) : filteredCatalogEntries.length === 0 ? (
                         <Card className="rounded-3xl p-0">
                             <CardContent className="flex h-[200px] items-center justify-center text-sm text-muted-foreground">
                                 No cards match the selected filters.
@@ -368,25 +513,25 @@ export default function CardsPage() {
                     ) : (
                         <>
                             <div className="grid gap-6 md:grid-cols-2">
-                                {pageItems.map((product) => {
-                                    const applied = appliedSlugs.has(product.slug ?? "") || appliedMatcher(product)
-                                    return (
-                                        <CatalogCreditCard
-                                            key={product.slug ?? product.product_name}
-                                            product={product}
-                                            applied={applied}
-                                            onApply={() => onApply(product)}
-                                        />
-                                    )
-                                })}
+                                {pageItems.map(({ product, applied }) => (
+                                    <CatalogCreditCard
+                                        key={product.slug ?? product.product_name}
+                                        product={product}
+                                        applied={applied}
+                                        isApplying={
+                                            pendingSlug === product.slug && (applyMutation.isPending || approveMutation.isPending)
+                                        }
+                                        onApply={() => onApply(product)}
+                                    />
+                                ))}
                             </div>
 
                             {/* Pagination: shows only when more than 4 results */}
-                            {filteredCatalog.length > PAGE_SIZE && (
+                            {filteredCatalogEntries.length > PAGE_SIZE && (
                                 <div className="flex flex-col items-center justify-between gap-3 sm:flex-row">
                                     <div className="text-xs text-muted-foreground">
-                                        Showing <b>{start + 1}</b>–<b>{Math.min(end, filteredCatalog.length)}</b> of{" "}
-                                        <b>{filteredCatalog.length}</b>
+                                        Showing <b>{start + 1}</b>–<b>{Math.min(end, filteredCatalogEntries.length)}</b> of{" "}
+                                        <b>{filteredCatalogEntries.length}</b>
                                     </div>
                                     <div className="flex items-center gap-2">
                                         <Button variant="outline" size="sm" onClick={() => setPage((p) => Math.max(1, p - 1))} disabled={page === 1}>
@@ -424,6 +569,11 @@ export default function CardsPage() {
 
 /* ===================== helpers ===================== */
 
+function isCardApplied(card: CardRowType): boolean {
+    const status = (card.status ?? "").toLowerCase()
+    return status === "applied" || Boolean(card.appliedAt)
+}
+
 function matchesAnnualFee(fee: number | null | undefined, filter: AnnualFeeFilter) {
     if (filter === "all") return true
     if (fee == null) return false
@@ -439,13 +589,6 @@ function formatAnnualFee(fee: number | null | undefined) {
     if (fee === 0) return "$0"
     return currency0.format(fee)
 }
-function formatForeignFee(value: number | null | undefined) {
-    if (value == null) return "None"
-    if (value <= 0) return "None"
-    if (value > 0 && value <= 1) return percent1.format(value)
-    return currency0.format(value)
-}
-
 function extractCatalogCards(raw: any): CreditCardProduct[] {
     if (!raw) return []
     if (Array.isArray(raw)) return raw
@@ -494,7 +637,7 @@ function buildAppliedMatcher(cards: CardRowType[]) {
         const joined = `${issuer} ${n1 || n2}`.trim()
         if (issuer && (n1 || n2)) issuerNameSet.add(joined)
 
-        const pslug = (c as any).productSlug
+        const pslug = c.cardProductSlug
         if (pslug) slugSet.add(pslug)
     }
 
@@ -502,20 +645,51 @@ function buildAppliedMatcher(cards: CardRowType[]) {
         const pName = norm(p.product_name)
         const pIssuer = norm(p.issuer)
         const pJoined = `${pIssuer} ${pName}`.trim()
-        return (
-            (p.slug && slugSet.has(p.slug)) ||
-            (pName && nameSet.has(pName)) ||
-            (pIssuer && pName && issuerNameSet.has(pJoined))
-        )
+        const matchesSlug = p.slug ? slugSet.has(p.slug) : false
+        const matchesName = pName ? nameSet.has(pName) : false
+        const matchesIssuerName = pIssuer && pName ? issuerNameSet.has(pJoined) : false
+        return matchesSlug || matchesName || matchesIssuerName
     }
 }
 
 /* ===================== glossy catalog card ===================== */
 
+type StatusFilterPillsProps = {
+    value: StatusFilter
+    onChange: (value: StatusFilter) => void
+}
+
+function StatusFilterPills({ value, onChange }: StatusFilterPillsProps) {
+    return (
+        <div className="flex items-center gap-1 rounded-full border border-border/60 bg-muted/50 p-0.5 text-xs shadow-sm">
+            {STATUS_FILTERS.map((option) => {
+                const isActive = option.value === value
+                return (
+                    <button
+                        key={option.value}
+                        type="button"
+                        onClick={() => onChange(option.value)}
+                        aria-pressed={isActive}
+                        className={[
+                            "rounded-full px-3 py-1 text-xs font-medium transition focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary/40",
+                            isActive
+                                ? "bg-primary text-primary-foreground shadow-sm"
+                                : "text-muted-foreground hover:text-foreground",
+                        ].join(" ")}
+                    >
+                        {option.label}
+                    </button>
+                )
+            })}
+        </div>
+    )
+}
+
 type CatalogCreditCardProps = {
     product: CreditCardProduct
     applied: boolean
     onApply: () => void
+    isApplying?: boolean
 }
 
 function gradientFor(product: CreditCardProduct) {
@@ -527,7 +701,7 @@ function gradientFor(product: CreditCardProduct) {
     return "from-violet-500 via-purple-500 to-fuchsia-600"
 }
 
-function CatalogCreditCard({ product, applied, onApply }: CatalogCreditCardProps) {
+function CatalogCreditCard({ product, applied, onApply, isApplying = false }: CatalogCreditCardProps) {
     const gradient = gradientFor(product)
     const issuer = (product.issuer ?? "").toUpperCase()
     const name = product.product_name
@@ -586,8 +760,15 @@ function CatalogCreditCard({ product, applied, onApply }: CatalogCreditCardProps
                             Applied
                         </Button>
                     ) : (
-                        <Button size="sm" onClick={onApply}>
-                            Apply
+                        <Button size="sm" onClick={onApply} disabled={isApplying}>
+                            {isApplying ? (
+                                <>
+                                    <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+                                    Applying…
+                                </>
+                            ) : (
+                                "Apply"
+                            )}
                         </Button>
                     )}
                     {product.link_url ? (
