+++ server/core/security.py
@@ -1,0 +1,57 @@
+"""Security helpers for decoding JWT tokens."""
+
+from typing import Any, Dict
+
+import requests
+from jose import jwt
+from jose.exceptions import JWTError
+from werkzeug.exceptions import Unauthorized
+
+JWKS_CACHE: Dict[str, Any] = {"keys": []}
+
+
+def fetch_jwks(jwks_url: str) -> Dict[str, Any]:
+    response = requests.get(jwks_url, timeout=5)
+    response.raise_for_status()
+    return response.json()
+
+
+def get_jwks(jwks_url: str) -> Dict[str, Any]:
+    if not JWKS_CACHE["keys"]:
+        JWKS_CACHE.update(fetch_jwks(jwks_url))
+    return JWKS_CACHE
+
+
+def get_rsa_key(token: str, jwks: Dict[str, Any]) -> Dict[str, Any]:
+    unverified_header = jwt.get_unverified_header(token)
+    for key in jwks.get("keys", []):
+        if key.get("kid") == unverified_header.get("kid"):
+            return {
+                "kty": key.get("kty"),
+                "kid": key.get("kid"),
+                "use": key.get("use"),
+                "n": key.get("n"),
+                "e": key.get("e"),
+            }
+    raise Unauthorized("Unable to find appropriate key")
+
+
+def decode_token(settings: Dict[str, str]) -> Dict[str, Any]:
+    from flask import request  # Imported lazily to avoid circular imports
+
+    auth_header = request.headers.get("Authorization", "")
+    if not auth_header.lower().startswith("bearer "):
+        raise Unauthorized("Authorization header must start with Bearer")
+    token = auth_header.split()[1]
+    jwks = get_jwks(settings["jwks_url"])
+    rsa_key = get_rsa_key(token, jwks)
+    try:
+        return jwt.decode(
+            token,
+            rsa_key,
+            algorithms=["RS256"],
+            audience=settings["audience"],
+            issuer=settings["issuer"],
+        )
+    except JWTError as exc:  # pragma: no cover - runtime validation
+        raise Unauthorized(f"Token verification failed: {exc}") from exc
